// src/display-api-client/api.ts
export interface CollectionJsonld {
    /**
     * 
     * @type {CollectionJsonldContext}
     * @memberof CollectionJsonld
     */
    '@context'?: CollectionJsonldContext;
    /**
     * 
     * @type {string}
     * @memberof CollectionJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionJsonld
     */
    '@type'?: string;
    /**
     * Checks whether the collection is empty (contains no elements).
     * @type {boolean}
     * @memberof CollectionJsonld
     */
    'empty'?: boolean;
    /**
     * Gets all keys/indices of the collection.
     * @type {Array<number>}
     * @memberof CollectionJsonld
     */
    'keys'?: Array<number>;
    /**
     * Gets all values of the collection.
     * @type {Array<string>}
     * @memberof CollectionJsonld
     */
    'values'?: Array<string>;
    /**
     * 
     * @type {any}
     * @memberof CollectionJsonld
     */
    'iterator'?: any;
}
/**
 * @type CollectionJsonldContext
 * @export
 */
export type CollectionJsonldContext = CollectionJsonldContextOneOf | string;

/**
 * 
 * @export
 * @interface CollectionJsonldContextOneOf
 */
export interface CollectionJsonldContextOneOf {
    [key: string]: any;

    /**
     * 
     * @type {string}
     * @memberof CollectionJsonldContextOneOf
     */
    '@vocab': string;
    /**
     * 
     * @type {string}
     * @memberof CollectionJsonldContextOneOf
     */
    'hydra': CollectionJsonldContextOneOfHydraEnum;
}
// ...
export interface PlaylistPlaylistJsonld {
    /**
     * 
     * @type {CollectionJsonldContext}
     * @memberof PlaylistPlaylistJsonld
     */
    '@context'?: CollectionJsonldContext;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    '@id'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    '@type'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'description'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof PlaylistPlaylistJsonld
     */
    'schedules'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'modified'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'modifiedBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'createdBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof PlaylistPlaylistJsonld
     */
    'slides'?: string;
    /**
     * 
     * @type {CollectionJsonld}
     * @memberof PlaylistPlaylistJsonld
     */
    'campaignScreens'?: CollectionJsonld;
    /**
     * 
     * @type {CollectionJsonld}
     * @memberof PlaylistPlaylistJsonld
     */
    'campaignScreenGroups'?: CollectionJsonld;
    /**
     * 
     * @type {CollectionJsonld}
     * @memberof PlaylistPlaylistJsonld
     */
    'tenants'?: CollectionJsonld;
    /**
     * 
     * @type {boolean}
     * @memberof PlaylistPlaylistJsonld
     */
    'isCampaign'?: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof PlaylistPlaylistJsonld
     */
    'published'?: Array<string>;
}
// ...
export interface PutV1PlaylistSlideIdRequestInner {
    /**
     * Slide ULID
     * @type {string}
     * @memberof PutV1PlaylistSlideIdRequestInner
     */
    'slide'?: string;
    /**
     * 
     * @type {number}
     * @memberof PutV1PlaylistSlideIdRequestInner
     */
    'weight'?: number;
}